Refactore large tests suite
