+++ aw-adsb/adsb-core/src/test/scala/io/syspulse/aeroware/adsb/core/DecoderSpec.scala
@@ -39,12 +39,6 @@ class DecoderSpec extends WordSpec with Matchers with Testables {
       a1.get.aircraftAddr should === (AircraftAddress("4840d6","Fokker 70","PH-KZD"))
     }
   
-    s"decode ${msg1} as ADSB_AircraftIdentification type" in {
-      val a1 = Decoder.decode(msg1)
-      a1.get.getClass should === (ADSB_AircraftIdentification(17,5,AircraftAddress("4840d6","",""),0,0,"",null).getClass)
-    }
-
-
     s"decode all ${file1} without crash" in {
       load(file1).map( message => {
         val a1 = Decoder.decode(message)
@@ -118,21 +112,6 @@ class DecoderSpec extends WordSpec with Matchers with Testables {
     s"decode 30 (invalid) to '#'" in {
       val v = Decoder.decodeCharacter(BitVector.fromBin(30.toBinaryString).get)
       v should === ('#')
-    }
-
-    s"decode '001011 001100 001101 110001 110000 110010 110011 100000' to 'KLM1023'" in {
-      val v = Decoder.decodeDataAsChars(Seq(bin"001011",bin"001100",bin"001101",bin"110001",bin"110000",bin"110010",bin"110011",bin"100000"))
-      v should === ("KLM1023")
-    }
-
-    s"decode ${msgIdent1} as ADSB_AircraftIdentification(MSI6215)" in {
-      val a1 = Decoder.decode(msgIdent1)
-      a1 should === (Success(ADSB_AircraftIdentification(17,5,AircraftAddress("508095","Antonov An-140","UR-14005"),4,0,"MSI6215",msgIdent1)))
-    }
-
-    s"decode ${msgIdent2} as ADSB_AircraftIdentification(KLM1023)" in {
-      val a1 = Decoder.decode(msgIdent2)
-      a1 should === (Success(ADSB_AircraftIdentification(17,5,AircraftAddress("4840D6","Fokker 70","PH-KZD"),4,0,"KLM1023",msgIdent2)))
     }
   }  
 }
